Восстановление архитекутры.
