+++ SecurityDoors.Tests/Implementations/TCPTest.cs
@@ -6,9 +6,8 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SecurityDoorModeling.Tests;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace SecurityDoorsModelingUnitTests
+namespace SecurityDoors.Tests.Implementations
 {
-   
     [TestClass]
     public class TCPTest
     {
