+++ SecurityDoors.Tests/SecurityDoors.Tests.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{CB2234CC-1FFE-4C47-8E0B-C11ABC1B053F}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>SecurityDoorsModelingUnitTests</RootNamespace>
-    <AssemblyName>SecurityDoorsModelingUnitTests</AssemblyName>
+    <RootNamespace>SecurityDoors.Tests</RootNamespace>
+    <AssemblyName>SecurityDoors.Tests</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
@@ -62,13 +62,13 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ConnectionSettingsTest.cs" />
-    <Compile Include="DataOperationsTest.cs" />
+    <Compile Include="Implementations\ConnectionSettingsTest.cs" />
+    <Compile Include="Implementations\DataOperationsTest.cs" />
     <Compile Include="ServiceFixture.cs" />
-    <Compile Include="Tests\CacheTest.cs" />
+    <Compile Include="Implementations\CacheTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TCPTest.cs" />
-    <Compile Include="WebConnectionTest.cs" />
+    <Compile Include="Implementations\TCPTest.cs" />
+    <Compile Include="Implementations\WebConnectionTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
